% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{actuals}
\alias{actuals}
\title{Weather Actuals for BPA}
\format{date column, unused column, actuals column \describe{
  \item{date}{normal datetime} \item{actuals}{actual wind generation value in mW} ... }}
\source{
The BPA releases
}
\usage{
actuals
}
\description{
A dataset containing the actuals for 12/2015 -> 05/2017 They are used in the
getVariogramScore function to compute the hourly wind correlations.
}
\keyword{datasets}
