% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_scripts.R
\name{getEnergyScore}
\alias{getEnergyScore}
\title{Title}
\usage{
getEnergyScore(type = 1, dataFrame = NULL, pyspDirPath, upperBound = 0)
}
\arguments{
\item{type}{Type 1 means we are reading a dataframe in the output format specified in prescient 1.0. Type 2 read in a pysp directory and get the info and make the dataframe}

\item{dataFrame}{If we use type 1, this is the R dataframe. Format is Prescient 1.0 output format.}

\item{pyspDirPath}{IF we are using type 2, this is the fullpath to the directory which contains the scenarios, actuals, and forecasts}

\item{upperBound}{The upper bound for scaling the scores}
}
\value{
Returns the energy score for the given day.
}
\description{
Title
}
