% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_scripts.R
\name{getVariogramScore}
\alias{getVariogramScore}
\title{Return Variogram score from a day's worth of scenarios}
\usage{
getVariogramScore(type = 1, dataFrame = NULL, pyspDirPath, p,
  upperBound = 0)
}
\arguments{
\item{type}{Type 1 means we are reading a dataframe in the output format specified in prescient 1.0. Type 2 read in a pysp directory and get the info and make the dataframe}

\item{dataFrame}{If type 1 is specified, this is the dataframe object. Not required for type 2}

\item{pyspDirPath}{If type 2 is specified, this is the path to the directory}

\item{p}{P paramter as specified by the variogram score documentation}

\item{upperBound}{If scale is true, this is the upper bound we are using to scale the data.}
}
\value{
Variogram score as a number.
}
\description{
Return Variogram score from a day's worth of scenarios
}
