% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_scripts.R
\name{plotScenario}
\alias{plotScenario}
\title{Plot Prescient scenario}
\usage{
plotScenario(type = 1, dataFrame = NULL, pyspDirPath = NULL, plot = T,
  saveID = FALSE, upperBound = 4782, plotFileName = "scenario_plot.pdf",
  title = "default title")
}
\arguments{
\item{type}{Type 1 indicates we are taking the data from a dataframe in the
prescient output format. Type 2 indicates we are pointing the function at a
path to the pyspdir output directory}

\item{dataFrame}{If type 1 is slected, this is the dataframe. The structure
should be the same as the Prescient output format}

\item{pyspDirPath}{If type 2 is selected, this is required as the path to the
specific date's folder in the pyspdir_2stage directory}

\item{plot}{T or F. Do we plot?}

\item{saveID}{T or F, do we save?}

\item{upperBound}{What do we use to scale the data by? Note: this has to be
adjusted if we change the wind scaling factors}

\item{plotFileName}{If we are plotting, what's our file name? Saved in the
current wd unless modified by a relative path.}

\item{title}{Title of the plot}
}
\value{
One kickass scenario plot
}
\description{
Plot Prescient scenario
}
